# swagger.json
{"openapi":"3.0.0","paths":{"/{network}/token":{"get":{"operationId":"TokenController_getTokens","summary":"Get a paginated list of tokens, optionally filter by type.","parameters":[{"name":"network","required":true,"in":"path","description":"Blockchain network name or chain id, currently only eth-mainnet (or 1) is supported","example":"eth-mainnet","schema":{"type":"string"}},{"name":"tokenType","required":false,"in":"query","description":"Filter by token type (erc20, erc721, erc777, erc1155)","example":"erc20","schema":{"enum":["erc20","erc721","erc777","erc1155"],"type":"string"}},{"name":"page","required":false,"in":"query","description":"Page number for pagination","example":1,"schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of tokens per page, Max 100 per page","example":10,"schema":{"type":"number"}}],"responses":{"200":{"description":"The paginated list of tokens.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TokenDto"}}}}}},"tags":["Tokens"]}},"/{network}/token/{contractAddress}":{"get":{"operationId":"TokenController_getTokenByAddress","summary":"Get token by contract address.","parameters":[{"name":"network","required":true,"in":"path","description":"Blockchain network name or chain id, currently only eth-mainnet (or 1) is supported","example":"eth-mainnet","schema":{"type":"string"}},{"name":"contractAddress","required":true,"in":"path","description":"The contract address of the token in hexadecimal format or an ENS name.","example":"0xCeD4E93198734dDaFf8492d525Bd258D49eb388E","schema":{"type":"string"}}],"responses":{"200":{"description":"The token details.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenDto"}}}}},"tags":["Tokens"]}},"/{network}/balance/single":{"get":{"operationId":"BalanceController_getSingleBalance","summary":"Get balance by wallet, token, and block number or timestamp.","parameters":[{"name":"network","required":true,"in":"path","description":"Blockchain network name or chain id, currently only eth-mainnet (or 1) is supported","example":"eth-mainnet","schema":{"type":"string"}},{"name":"holder","required":false,"in":"query","description":"The wallet address or ENS name of the user.","example":"0x1234567890abcdef1234567890abcdef12345678","schema":{"type":"string"}},{"name":"tokenAddress","required":false,"in":"query","description":"The contract address of the token.","example":"0xabcdefabcdefabcdefabcdefabcdefabcdefabcd","schema":{"type":"string"}},{"name":"blockNumber","required":false,"in":"query","description":"Optional block number for which the balance is requested.","example":"123456","schema":{"type":"number"}},{"name":"timestamp","required":false,"in":"query","description":"Optional timestamp for which the balance is requested (ISO 8601 format).","example":"2022-01-01T00:00:00Z","schema":{"type":"string"}}],"responses":{"200":{"description":"The balance data for a specific block or timestamp.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BalanceDto"}}}}},"tags":["Balances"]}},"/{network}/balance/range":{"get":{"operationId":"BalanceController_getRangeBalances","summary":"Get balances over a block range or time period for a given network.","parameters":[{"name":"version","required":true,"in":"path","schema":{"type":"string"}},{"name":"network","required":true,"in":"path","description":"Blockchain network name or chain id, currently only eth-mainnet (or 1) is supported","example":"eth-mainnet","schema":{"type":"string"}},{"name":"holder","required":false,"in":"query","description":"The wallet address or ENS name of the user.","example":"0x1234567890abcdef1234567890abcdef12345678","schema":{"type":"string"}},{"name":"tokenAddress","required":false,"in":"query","description":"The contract address of the token.","example":"0xabcdefabcdefabcdefabcdefabcdefabcdefabcd","schema":{"type":"string"}},{"description":"Start of the block range or timestamp (ISO 8601).","example":"2022-01-01T00:00:00Z","name":"start","in":"query","required":true,"schema":{"type":"string"}},{"description":"End of the block range or timestamp (ISO 8601).","example":"2022-01-02T00:00:00Z","name":"end","in":"query","required":true,"schema":{"type":"string"}},{"name":"startBlock","required":false,"in":"query","description":"Start block number for querying balances over time.","example":123456,"schema":{"type":"number"}},{"name":"endBlock","required":false,"in":"query","description":"End block number for querying balances over time.","example":123999,"schema":{"type":"number"}}],"responses":{"200":{"description":"The balances data for a range of blocks or timestamps.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BalanceDto"}}}}}},"tags":["Balances"]}},"/{network}/allowance/single":{"get":{"operationId":"AllowanceController_getAllowance","summary":"Get allowance by owner, spender, token, and block number for a given network.","parameters":[{"name":"network","required":true,"in":"path","description":"Blockchain network name or chain id, currently only eth-mainnet (or 1) is supported","example":"eth-mainnet","schema":{"type":"string"}},{"name":"owner","required":false,"in":"query","description":"The wallet address or ENS name of the owner.","example":"0x1234567890abcdef1234567890abcdef12345678","schema":{"type":"string"}},{"name":"spender","required":false,"in":"query","description":"The wallet address or ENS name of the spender.","example":"0xabcdefabcdefabcdefabcdefabcdefabcdefabcd","schema":{"type":"string"}},{"name":"tokenAddress","required":false,"in":"query","description":"The contract address of the token.","example":"0xabcdefabcdefabcdefabcdefabcdefabcdefabcd","schema":{"type":"string"}},{"name":"blockNumber","required":false,"in":"query","description":"The block number for which the allowance is requested.","example":"123456","schema":{"type":"number"}},{"name":"timestamp","required":false,"in":"query","description":"The timestamp for which the allowance is requested (ISO 8601 format).","example":"2022-01-01T00:00:00Z","schema":{"type":"string"}}],"responses":{"200":{"description":"The allowance data.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AllowanceDto"}}}}},"tags":["Allowances"]}},"/{network}/allowance/range":{"get":{"operationId":"AllowanceController_getRangeBalances","summary":"Get allowances over a block range or time period for a given network.","parameters":[{"name":"network","required":true,"in":"path","description":"Blockchain network name or chain id, currently only eth-mainnet (or 1) is supported","example":"eth-mainnet","schema":{"type":"string"}},{"name":"owner","required":false,"in":"query","description":"The wallet address or ENS name of the owner.","example":"0x1234567890abcdef1234567890abcdef12345678","schema":{"type":"string"}},{"name":"spender","required":false,"in":"query","description":"The wallet address or ENS name of the spender.","example":"0xabcdefabcdefabcdefabcdefabcdefabcdefabcd","schema":{"type":"string"}},{"name":"tokenAddress","required":false,"in":"query","description":"The contract address of the token.","example":"0xabcdefabcdefabcdefabcdefabcdefabcdefabcd","schema":{"type":"string"}},{"description":"Start of the block range or timestamp (ISO 8601).","example":"2022-01-01T00:00:00Z","required":false,"name":"start","in":"query","schema":{"type":"string"}},{"description":"End of the block range or timestamp (ISO 8601).","example":"2022-01-02T00:00:00Z","required":false,"name":"end","in":"query","schema":{"type":"string"}},{"name":"startBlock","required":false,"in":"query","description":"Start block number for querying allowances over time.","example":123456,"schema":{"type":"number"}},{"name":"endBlock","required":false,"in":"query","description":"End block number for querying allowances over time.","example":123999,"schema":{"type":"number"}}],"responses":{"200":{"description":"The allowances data for a range of blocks or timestamps.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AllowanceDto"}}}}}},"tags":["Allowances"]}},"/{network}/token-supply":{"get":{"operationId":"TokenSupplyController_getTokenSupply","summary":"Get token supply by token address and block number for a given network.","parameters":[{"name":"network","required":true,"in":"path","description":"Blockchain network name or chain id, currently only eth-mainnet (or 1) is supported","example":"eth-mainnet","schema":{"type":"string"}},{"name":"tokenAddress","required":true,"in":"query","description":"The contract address of the token or ENS name.","example":"0xabcdefabcdefabcdefabcdefabcdefabcdefabcd","schema":{"type":"string"}},{"name":"blockNumber","required":true,"in":"query","description":"The block number for which the token supply is requested.","example":"123456","schema":{"type":"string"}}],"responses":{"200":{"description":"The token supply data.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenSupplyDto"}}}}},"tags":["TokenSupplies"]}},"/{network}/nft/token-uri":{"get":{"operationId":"NFTController_getTokenUri","summary":"Get token URI for an ERC721 or ERC1155 token","parameters":[{"name":"network","required":true,"in":"path","description":"Blockchain network name or chain id, currently only eth-mainnet (or 1) is supported","example":"eth-mainnet","schema":{"type":"string"}},{"name":"tokenAddress","required":true,"in":"path","description":"The contract address of the ERC721 or ERC1155 token.","example":"0xabcdefabcdefabcdefabcdefabcdefabcdefabcd","schema":{"type":"string"}},{"name":"tokenId","required":true,"in":"path","description":"The token ID for which the token URI is requested.","example":"1","schema":{"type":"string"}}],"responses":{"200":{"description":"The token URI."},"400":{"description":"Invalid contract or token ID."}},"tags":["NFT"]}},"/{network}/nft/owner":{"get":{"operationId":"NFTController_isOwnerOfToken","summary":"Check if the user is the owner of an NFT. Works for both ERC721 and ERC1155","parameters":[{"name":"network","required":true,"in":"path","description":"Blockchain network name or chain id, currently only eth-mainnet (or 1) is supported","example":"eth-mainnet","schema":{"type":"string"}},{"name":"tokenAddress","required":true,"in":"path","description":"The contract address of the ERC721 or ERC1155 token.","example":"0xabcdefabcdefabcdefabcdefabcdefabcdefabcd","schema":{"type":"string"}},{"name":"owner","required":true,"in":"path","description":"The ENS name or address of the owner.","schema":{"type":"string"}},{"name":"tokenId","required":true,"in":"path","description":"The token ID for which the token URI is requested.","example":"1","schema":{"type":"string"}}],"responses":{"200":{"description":"Boolean indicating if the user is the owner of the NFT."},"400":{"description":"Invalid contract or token ID."}},"tags":["NFT"]}},"/{network}/common/block-height":{"get":{"operationId":"CommonController_getBlockHeight","summary":"Get the current block height","parameters":[{"name":"network","required":true,"in":"path","description":"Blockchain network name or chain id, currently only eth-mainnet (or 1) is supported","example":"eth-mainnet","schema":{"type":"string"}}],"responses":{"200":{"description":"The current block height."},"400":{"description":"Failed to fetch block height."}},"tags":["Common"]}},"/{network}/common/gas-price":{"get":{"operationId":"CommonController_getGasPrice","summary":"Get the current gas price","parameters":[{"name":"network","required":true,"in":"path","description":"Blockchain network name or chain id, currently only eth-mainnet (or 1) is supported","example":"eth-mainnet","schema":{"type":"string"}},{"name":"type","required":true,"in":"path","description":"Type of transaction: native token transfer, ERC20 transfer, or swap.","schema":{"enum":["native_transfer","erc20_transfer","swap"],"type":"string"}}],"responses":{"200":{"description":"The current gas price in Gwei."},"400":{"description":"Failed to fetch gas price."}},"tags":["Common"]}},"/{network}/common/logs":{"get":{"operationId":"CommonController_getLogs","summary":"Get logs for a contract address","parameters":[{"name":"network","required":true,"in":"path","description":"Blockchain network name or chain id, currently only eth-mainnet (or 1) is supported","example":"eth-mainnet","schema":{"type":"string"}},{"name":"startBlock","required":false,"in":"query","description":"The starting block number.","example":123456,"schema":{"type":"number"}},{"name":"endBlock","required":false,"in":"query","description":"The ending block number.","example":123999,"schema":{"type":"number"}},{"name":"blockHash","required":false,"in":"query","description":"The block hash to filter logs.","example":"0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef","schema":{"type":"string"}},{"name":"contractAddress","required":false,"in":"query","description":"The contract address to filter logs.","example":"0xabcdefabcdefabcdefabcdefabcdefabcdefabcd","schema":{"type":"string"}},{"name":"topics","required":false,"in":"query","description":"The topics array to filter logs.","example":["0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"],"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"The logs for the specified contract."},"400":{"description":"Failed to fetch logs."}},"tags":["Common"]}},"/{network}/common/block":{"get":{"operationId":"CommonController_getBlock","summary":"Get a block by block number or block hash on any supported network","parameters":[{"name":"network","required":true,"in":"path","description":"Blockchain network name or chain id, currently only eth-mainnet (or 1) is supported","example":"eth-mainnet","schema":{"type":"string"}},{"name":"blockNumber","required":false,"in":"path","description":"The block number for which the block data is requested.","example":"123456","schema":{"type":"string"}},{"name":"blockHash","required":false,"in":"path","description":"The block hash for which the block data is requested.","example":"0xabcdefabcdefabcdefabcdefabcdefabcdefabcd","schema":{"type":"string"}}],"responses":{"200":{"description":"The block details."},"400":{"description":"Failed to fetch block."}},"tags":["Common"]}},"/{network}/transactions/{txHash}":{"get":{"operationId":"TransactionController_getTransactionDetails","summary":"Fetch transaction details by hash","parameters":[{"name":"network","required":true,"in":"path","description":"Blockchain network name or chain id, currently only eth-mainnet (or 1) is supported","example":"eth-mainnet","schema":{"type":"string"}},{"name":"txHash","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Transaction details fetched successfully."},"400":{"description":"Invalid network name or transaction hash."}},"tags":["Transactions"]}},"/{network}/uniswap/eth-usdt-price":{"get":{"operationId":"UniswapV3Controller_getETHToUSDTPrice","summary":"Get ETH to USDT price from Uniswap V3 pool on a specified network","parameters":[{"name":"network","required":true,"in":"path","description":"Blockchain network name or chain id, currently only eth-mainnet (or 1) is supported","example":"eth-mainnet","schema":{"type":"string"}}],"responses":{"200":{"description":"The ETH/USDT price."},"400":{"description":"Invalid network name."}},"tags":["UniswapV3"]}}},"info":{"title":"Histori API","description":"API for token data and historical balances","version":"0.0.5","contact":{}},"tags":[],"servers":[{"url":"https://api.histori.xyz","description":"Production server"},{"url":"http://localhost:3001","description":"Local development server"}],"components":{"securitySchemes":{"apiKey":{"type":"apiKey","in":"header","name":"x-api-key"}},"schemas":{"TokenDto":{"type":"object","properties":{"contractAddress":{"type":"string","description":"Token contract address in hexadecimal format","example":"0xabcdefabcdefabcdefabcdefabcdefabcdefabcd"},"blockNumber":{"type":"number","description":"Block number at the time of token creation","example":123456},"tokenType":{"type":"string","description":"Type of the token (e.g., ERC20, ERC721, ERC777, ERC1155)","example":"ERC20"},"name":{"type":"string","description":"Token name (required for all token types)","example":"My Token"},"symbol":{"type":"string","description":"Token symbol (required for all token types)","example":"MTK"},"decimals":{"type":"number","description":"Number of decimals for ERC20 tokens","example":18},"granularity":{"type":"string","description":"Granularity for ERC777 tokens","example":"1","nullable":true}},"required":["contractAddress","blockNumber","tokenType","name","symbol","decimals","granularity"]},"BalanceDto":{"type":"object","properties":{"holder":{"type":"string","description":"Wallet address of the token holder in hexadecimal format","example":"0x1234567890abcdef1234567890abcdef12345678"},"contractAddress":{"type":"string","description":"Token contract address in hexadecimal format","example":"0xabcdefabcdefabcdefabcdefabcdefabcdefabcd"},"balance":{"type":"string","description":"The balance of the token","example":"1000000000000000000"},"blockNumber":{"type":"number","description":"Block number at the time the balance was recorded","example":123456},"tokenId":{"type":"string","description":"Token ID for ERC721 tokens (optional)","example":1,"nullable":true}},"required":["holder","contractAddress","balance","blockNumber","tokenId"]},"AllowanceDto":{"type":"object","properties":{"contractAddress":{"type":"string","description":"The contract address of the token in hexadecimal format."},"owner":{"type":"string","description":"The ENS name or address of the owner."},"spender":{"type":"string","description":"The ENS name or address of the spender."},"allowance":{"type":"string","description":"The allowance amount. Optional for ERC721 and ERC1155 tokens, which do not have a specific allowance value."},"blockNumber":{"type":"number","description":"The block number for which the allowance is requested."},"tokenType":{"type":"string","description":"The type of token: erc20, erc721, erc777, or erc1155."},"tokenId":{"type":"string","description":"The token ID, used for non-fungible tokens like ERC721 and ERC1155. Optional."}},"required":["contractAddress","owner","spender","blockNumber","tokenType"]},"TokenSupplyDto":{"type":"object","properties":{"contractAddress":{"type":"string","description":"Token contract address in hexadecimal format","example":"0xabcdefabcdefabcdefabcdefabcdefabcdefabcd"},"blockNumber":{"type":"number","description":"Block number at the time of the snapshot","example":123456},"totalSupply":{"type":"string","description":"Total supply of the token at the given block number","example":"1000000000000000000"}},"required":["contractAddress","blockNumber","totalSupply"]}}},"security":[{"apiKey":[]}]}